requirements.txt
Django
django-crispy-forms
crispy-bootstrap4
pillow